Move model types to model modules
