+++ app/models/unit.ts
@@ -1,6 +1,8 @@
 import { z } from "zod";
 import isHexColor from "validator/lib/isHexColor";
 
+export type { Unit } from "@prisma/client";
+
 const SHORT_TEXT_LIMIT = 255;
 const LONG_TEXT_LIMIT = SHORT_TEXT_LIMIT * 4;
 const MIN_REQUIRED_TEXT = 1;
