+++ app/routes/_auth.armies.new.tsx
@@ -10,10 +10,11 @@ import {
 import * as R from "ramda";
 import invariant from "tiny-invariant";
 import { ZodError } from "zod";
-import db, { Army } from "~/.server/db";
 
+import db from "~/.server/db";
 import { Button, FormField, PageHeading } from "~/components";
 import { formatValidationErrors } from "~/utils/form";
+import { Army } from "~/models/army";
 
 type FormErrors = Partial<Record<keyof Army, string[]>>;
